fab file now finally working properly

+++ applications/tpcc/src/main/java/com/codefutures/tpcc/Driver.java
@@ -45,7 +45,7 @@ public class Driver implements TpccConstants {
     private int[][] late2;
     private int[][] retry2;
     private int[][] failure2;
-	private int[] latencies;
+	private double[] latencies;
 
     public double[] max_rt = new double[TRANSACTION_COUNT];
 
@@ -74,7 +74,8 @@ public class Driver implements TpccConstants {
      */
     public Driver(Connection conn, int fetchSize,
                   int[] success, int[] late, int[] retry, int[] failure,
-                  int[][] success2, int[][] late2, int[][] retry2, int[][] failure2, int[] latencies, boolean joins) {
+                  int[][] success2, int[][] late2, int[][] retry2, int[][] failure2, double[] latencies, boolean
+						  joins) {
         try {
 			this.conn = conn;
